@@ -9,7 +9,7 @@
 		}
 	],
 	"require": {
-		"php": ">=7.3",
+		"php": ">=7.4",
 		"monolog/monolog": "^2.0",
 		"psr/container": "1.0.0"
 	},
@@ -18,9 +18,16 @@
 	},
 	"require-dev": {
 		"ergebnis/composer-normalize": "^2.5",
-		"friendsofphp/php-cs-fixer": "^2.11",
+		"friendsofphp/php-cs-fixer": "^2.11@stable",
+		"holonet/hdev": "~1.0.0",
 		"phpunit/phpunit": "^8.4.1",
 		"vimeo/psalm": "^3.5"
+	},
+	"extra": {
+		"branch-alias": {
+			"dev-master": "1.3.x-dev",
+			"dev-develop": "1.4.x-dev"
+		}
 	},
 	"autoload": {
 		"psr-4": {
@@ -35,6 +42,13 @@
 			"holonet\\common\\tests\\": "tests/"
 		}
 	},
+	"repositories": [
+		{
+			"type": "composer",
+			"url": "https://holonet.easylabs.ch/hgit/composer/"
+		}
+	],
+	"minimum-stability": "dev",
 	"scripts": {
 		"fix": [
 			"@composer normalize --diff",
