@@ -3,8 +3,6 @@
  * This file is part of the hdev common library package
  * (c) Matthias Lantsch.
  *
- * class file for the ComparableInterface interface
- *
  * @license http://www.wtfpl.net/ Do what the fuck you want Public License
  * @author  Matthias Lantsch <matthias.lantsch@bluewin.ch>
  */
@@ -12,7 +10,7 @@
 namespace holonet\common;
 
 /**
- * The IComparable interface forces the implementing class to define a comparison.
+ * The ComparableInterface interface forces the implementing class to define a comparison method.
  */
 interface ComparableInterface {
 	public function compareTo(self $other): bool;
