@@ -3,8 +3,6 @@
  * This file is part of the hdev common library package
  * (c) Matthias Lantsch.
  *
- * class file for the UnsupportedFormatException class
- *
  * @license http://www.wtfpl.net/ Do what the fuck you want Public License
  * @author  Matthias Lantsch <matthias.lantsch@bluewin.ch>
  */
@@ -12,14 +10,9 @@
 namespace holonet\common\config\exception;
 
 /**
- * Exception to be thrown when a specified file to read cannot be found.
+ * Exception to be thrown when a specified file to read is in a format the parser can't understand.
  */
 class UnsupportedFormatException extends ConfigReaderException {
-	/**
-	 * UnsupportedFormatException constructor.
-	 * Overwritten so we can submit the constant error code.
-	 * @param string $message String error message
-	 */
 	public function __construct(string $message) {
 		parent::__construct($message, static::UNSUPPORTED_FORMAT);
 	}
